requirements.txt
fastapi~=0.119.1
uvicorn[standard]
jinja2
pydantic~=2.12.3

paho-mqtt~=2.1.0